requirements.txt
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.1
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.0
branca==0.5.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.4
Django==4.0.4
django-cors-headers==3.12.0
djangorestframework==3.13.1
dnspython==2.2.1
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
filelock==3.6.0
folium==0.12.1.post1
idna==3.3
ipykernel==6.12.1
ipython==8.2.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
jedi==0.18.1
Jinja2==3.1.1
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.2.2
jupyter-console==6.4.3
jupyter-core==4.9.2
jupyterlab-pygments==0.2.0
jupyterlab-widgets==1.1.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.5.13
nbconvert==6.4.5
nbformat==5.3.0
nest-asyncio==1.5.5
notebook==6.4.10
numpy==1.22.4
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.1
pipenv==2022.4.8
platformdirs==2.5.1
prometheus-client==0.14.1
prompt-toolkit==3.0.29
psutil==5.9.0
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.11.2
pymongo==4.1.1
pyparsing==3.0.8
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.1
pyzmq==22.3.0
qtconsole==5.3.0
QtPy==2.0.1
requests==2.28.0
Send2Trash==1.8.0
six==1.16.0
soupsieve==2.3.2
sqlparse==0.4.2
stack-data==0.2.0
terminado==0.13.3
testpath==0.6.0
tornado==6.1
traitlets==5.1.1
urllib3==1.26.9
virtualenv==20.14.0
virtualenv-clone==0.5.7
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==6.1.0
widgetsnbextension==3.6.0
